requirements.txt
boto3==1.28.51
Flask==2.3.3
Flask-JWT-Extended==4.5.2
flask-marshmallow==0.15.0
Flask-RESTful==0.3.10
Flask-Reuploaded==1.3.0
flask-smorest==0.42.1
numpy==1.26.0
pandas==2.1.0
pyspark==3.4.1
python-dotenv==1.0.0
scikit-learn==1.3.0
uuid==1.30
typing==3.7.4.3